test: restructured tests for more crdts
