+++ tests/e2e/package.json
@@ -6,16 +6,16 @@
 		"dev": "concurrently \"pnpm:dev:client\" \"pnpm:dev:server:cloudflare\" \"pnpm:dev:server:node\" \"pnpm:dev:server:node-redis\"",
 		"dev:client": "next dev --port 3100",
 		"dev:server:cloudflare": "wrangler dev --port 3101",
-		"dev:server:node": "ts-node --project ./src/server/node/tsconfig.json ./src/server/node -- --port 3102",
-		"dev:server:node-redis": "ts-node --project ./src/server/node/tsconfig.json ./src/server/node-redis -- --port 3103",
+		"dev:server:node": "ts-node --project ./src/server/yjs/node/tsconfig.json ./src/server/yjs/node -- --port 3102",
+		"dev:server:node-redis": "ts-node --project ./src/server/yjs/node-redis/tsconfig.json ./src/server/yjs/node-redis -- --port 3103",
 		"migrate:dev": "prisma migrate dev",
 		"prisma": "prisma",
 		"playwright": "playwright",
 		"start": "next start",
 		"test": "playwright test",
-		"test:cloudflare": "wrangler deploy --dry-run && playwright test ./src/tests/cloudflare/**/*",
-		"test:node": "prisma migrate reset --force && playwright test ./src/tests/node/**/*",
-		"test:node-redis": "playwright test ./src/tests/node-redis/**/*",
+		"test:cloudflare": "wrangler deploy --dry-run && playwright test ./src/tests/yjs/cloudflare/**/*",
+		"test:node": "prisma migrate reset --force && playwright test ./src/tests/yjs/node/**/*",
+		"test:node-redis": "playwright test ./src/tests/yjs/node-redis/**/*",
 		"tsc:check": "tsc --noEmit"
 	},
 	"private": true,
