+++ tests/e2e/src/pages/cloudflare/storage/index.tsx
@@ -2,7 +2,7 @@ import { NextPage } from "next";
 import { useRouter } from "next/router";
 import { useState } from "react";
 import { CloudflareStorageRoom } from "../../../components";
-import { PluvRoomProvider } from "../../../pluv-io/cloudflare";
+import { PluvRoomProvider } from "../../../pluv-io/yjs/cloudflare";
 
 export const Page: NextPage = () => {
     const router = useRouter();
