+++ tests/e2e/src/pluv-io/yjs/node.ts
@@ -1,14 +1,15 @@
+import { addonIndexedDB } from "@pluv/addon-indexeddb";
 import { yjs } from "@pluv/crdt-yjs";
 import { createBundle, createClient } from "@pluv/react";
 import { z } from "zod";
-import type { io } from "../server/cloudflare";
+import type { io } from "../../server/yjs/node";
 
 const client = createClient<typeof io>({
     authEndpoint: (room) => {
-        return `http://localhost:3101/api/pluv/authorize?room=${room}`;
+        return `http://localhost:3102/api/pluv/authorize?room=${room}`;
     },
     wsEndpoint: (room) => {
-        return `ws://localhost:3101/api/pluv/room/${room}`;
+        return `ws://localhost:3102/api/pluv/room/${room}`;
     },
 });
 
@@ -43,6 +44,11 @@ export const {
     usePluvTransact,
     usePluvUndo,
 } = createRoomBundle({
+    addons: [
+        addonIndexedDB({
+            enabled: (room) => room.id === "e2e-node-storage-addon-indexeddb",
+        }),
+    ],
     initialStorage: yjs.doc(() => ({
         messages: yjs.array([
             yjs.object({
