--- tests/e2e/src/server/node-redis/tsconfig.json
@@ -1,21 +1,0 @@
-{
-	"extends": "@pluv/tsconfig/node.json",
-	"include": ["**/*.ts"],
-	"compilerOptions": {
-		"baseUrl": ".",
-		"declaration": false,
-		"declarationMap": false,
-		"types": ["@types/node"],
-		"paths": {
-			"@pluv/client": ["../../../../../packages/client/dist"],
-			"@pluv/crdt-yjs": ["../../../../../packages/crdt-yjs/dist"],
-			"@pluv/io": ["../../../../../packages/io/dist"],
-			"@pluv/persistance-redis": ["../../../../../packages/persistance-redis/dist"],
-			"@pluv/platform-node": ["../../../../../packages/platform-node/dist"],
-			"@pluv/pubsub-redis": ["../../../../../packages/pubsub-redis/dist"],
-			"@pluv/react": ["../../../../../packages/react/dist"],
-			"@pluv/types": ["../../../../../packages/types/dist"],
-			"ws": ["../../node_modules/@types/ws"]
-		}
-	}
-}
