+++ tests/e2e/src/server/yjs/cloudflare/tsconfig.json
@@ -1,0 +1,19 @@
+{
+	"extends": "@pluv/tsconfig/wrangler.json",
+	"include": [
+		"**/*.ts",
+		"**/*.tsx"
+	],
+	"compilerOptions": {
+		"baseUrl": ".",
+		"declaration": false,
+		"declarationMap": false,
+		"types": ["@cloudflare/workers-types"],
+		"paths": {
+			"@pluv-internal/cloudflare-utils": ["../../../../../../internal/cloudflare-utils/dist"],
+			"@pluv/io": ["../../../../../../packages/io/dist"],
+			"@pluv/platform-cloudflare": ["../../../../../../packages/platform-cloudflare/dist"],
+			"@pluv/types": ["../../../../../../packages/types/dist"]
+		}
+	}
+}
