+++ tests/e2e/src/server/yjs/node/pluv-io.ts
@@ -2,7 +2,7 @@ import { yjs } from "@pluv/crdt-yjs";
 import { createIO } from "@pluv/io";
 import { platformNode } from "@pluv/platform-node";
 import { z } from "zod";
-import { prisma } from "../../prisma";
+import { prisma } from "../../../prisma";
 
 const PLUV_AUTH_SECRET = "secret123";
