+++ tests/e2e/src/tests/yjs/cloudflare/presence/index.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "@playwright/test";
 import { oneLine } from "common-tags";
 import ms from "ms";
-import { openTestPage, waitMs } from "../../../utils";
+import { openTestPage, waitMs } from "../../../../utils";
 
 const TEST_URL = "http://localhost:3100/cloudflare/presence";
 
@@ -39,7 +39,7 @@ test.describe("Cloudflare Presence", () => {
 
             await firstPage.close();
             await secondPage.close();
-        }
+        },
     );
 
     test(
@@ -89,7 +89,7 @@ test.describe("Cloudflare Presence", () => {
 
             await firstPage.close();
             await secondPage.close();
-        }
+        },
     );
 
     test(
@@ -139,7 +139,7 @@ test.describe("Cloudflare Presence", () => {
 
             await firstPage.close();
             await secondPage.close();
-        }
+        },
     );
 
     test(
@@ -214,6 +214,6 @@ test.describe("Cloudflare Presence", () => {
 
             await firstPage.close();
             await secondPage.close();
-        }
+        },
     );
 });
