+++ tests/e2e/src/tests/yjs/node-redis/storage/index.test.ts
@@ -1,11 +1,11 @@
 import { expect, test } from "@playwright/test";
 import { oneLine } from "common-tags";
 import ms from "ms";
-import { openTestPage, waitMs } from "../../../utils";
+import { openTestPage, waitMs } from "../../../../utils";
 
-const TEST_URL = "http://localhost:3100/cloudflare/storage";
+const TEST_URL = "http://localhost:3100/node-redis/storage";
 
-test.describe("Cloudflare Storage", () => {
+test.describe("Node Redis Storage", () => {
     test(
         oneLine`
 			connect 1 ->
@@ -126,6 +126,148 @@ test.describe("Cloudflare Storage", () => {
             await waitMs(ms("0.25s"));
 
             const secondPage = await openTestPage(testUrl);
+
+            await secondPage.waitForSelector("#storage");
+            await waitMs(ms("0.25s"));
+
+            await secondPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(2));
+
+            await firstPage.click("#disconnect-room");
+            await waitMs(ms("0.25s"));
+
+            await secondPage.click("#button-add-message");
+            await waitMs(ms("0.25s"));
+
+            await firstPage.click("#connect-room");
+            await firstPage.waitForSelector("#storage-room");
+            await waitMs(ms("0.25s"));
+
+            await firstPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(3));
+
+            await firstPage.click("#disconnect-room");
+            await secondPage.click("#disconnect-room");
+            await waitMs(ms("0.25s"));
+
+            await firstPage.click("#connect-room");
+            await firstPage.waitForSelector("#storage-room");
+            await waitMs(ms("0.25s"));
+
+            await firstPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(1));
+
+            await firstPage.close();
+            await secondPage.close();
+        },
+    );
+
+    test(
+        oneLine`
+            connect 1 to pluv1 ->
+            connect 2 to pluv2 ->
+            verify storage on 1 + 2 ->
+            add message on 1 ->
+            verify storage on 1 + 2 ->
+            add mesage on 2 ->
+            verify storage on 1 + 2
+        `,
+        async () => {
+            const testUrl = `${TEST_URL}?room=e2e-storage-4`;
+
+            const firstPage = await openTestPage(testUrl);
+            const secondPage = await openTestPage(`${testUrl}_1`);
+
+            await Promise.all([
+                firstPage.waitForSelector("#storage"),
+                secondPage.waitForSelector("#storage"),
+            ]);
+
+            await waitMs(ms("0.25s"));
+
+            await firstPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(1));
+
+            await secondPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(1));
+
+            await firstPage.click("#button-add-message");
+            await waitMs(ms("0.25s"));
+
+            await firstPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(2));
+
+            await secondPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(2));
+
+            await secondPage.click("#button-add-message");
+            await waitMs(ms("0.25s"));
+
+            await firstPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(3));
+
+            await secondPage
+                .locator("#storage")
+                .innerText()
+                .then((text) => JSON.parse(text))
+                .then((messages) => expect(messages.length).toEqual(3));
+
+            await firstPage.close();
+            await secondPage.close();
+        },
+    );
+
+    test(
+        oneLine`
+            connect 1 to pluv1 ->
+            add message on 1 ->
+            connect 2 to pluv2 ->
+            verify storage on 2 ->
+            disconnect 1 ->
+            add message on 2 ->
+            connect 1 to pluv1 ->
+            verify storage on 1 ->
+            disconnect 1 ->
+            disconnect 2 ->
+            connect 1 ->
+            verify storage on 1 + 2
+        `,
+        async () => {
+            const testUrl = `${TEST_URL}?room=e2e-storage-5`;
+
+            const firstPage = await openTestPage(testUrl);
+
+            await firstPage.waitForSelector("#storage");
+            await waitMs(ms("0.25s"));
+
+            await firstPage.click("#button-add-message");
+            await waitMs(ms("0.25s"));
+
+            const secondPage = await openTestPage(`${testUrl}_1`);
 
             await secondPage.waitForSelector("#storage");
             await waitMs(ms("0.25s"));
